Implement `iced::Sandbox` trait for simple apps
